minour changes on the firebase
