+++ lib/firebase/firebase_options.dart
@@ -1,8 +1,85 @@
-import 'package:firebase_core/firebase_core.dart';
-import 'firebase_options.dart';
+// File generated by FlutterFire CLI.
+// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
+import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
+import 'package:flutter/foundation.dart'
+    show defaultTargetPlatform, kIsWeb, TargetPlatform;
 
-// ...
+/// Default [FirebaseOptions] for use with your Firebase apps.
+///
+/// Example:
+/// ```dart
+/// import 'firebase_options.dart';
+/// // ...
+/// await Firebase.initializeApp(
+///   options: DefaultFirebaseOptions.currentPlatform,
+/// );
+/// ```
+class DefaultFirebaseOptions {
+  static FirebaseOptions get currentPlatform {
+    if (kIsWeb) {
+      return web;
+    }
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+        return android;
+      case TargetPlatform.iOS:
+        return ios;
+      case TargetPlatform.macOS:
+        return macos;
+      case TargetPlatform.windows:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for windows - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      default:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions are not supported for this platform.',
+        );
+    }
+  }
 
-await Firebase.initializeApp(
-options: DefaultFirebaseOptions.currentPlatform,
-);
+  static const FirebaseOptions web = FirebaseOptions(
+    apiKey: 'AIzaSyDZCR-MSgjsZXQoruCL8siMaNouFVSIaxU',
+    appId: '1:667274875306:web:bfa4c86d168af5deaa9bd3',
+    messagingSenderId: '667274875306',
+    projectId: 'e-commerce-application-71a24',
+    authDomain: 'e-commerce-application-71a24.firebaseapp.com',
+    storageBucket: 'e-commerce-application-71a24.appspot.com',
+    measurementId: 'G-7DH8BSG22P',
+  );
+
+  static const FirebaseOptions android = FirebaseOptions(
+    apiKey: 'AIzaSyAK9p3H57DYBMmwWcZ9bR0bOFOUjeAOofM',
+    appId: '1:667274875306:android:3c50def0a4845690aa9bd3',
+    messagingSenderId: '667274875306',
+    projectId: 'e-commerce-application-71a24',
+    storageBucket: 'e-commerce-application-71a24.appspot.com',
+  );
+
+  static const FirebaseOptions ios = FirebaseOptions(
+    apiKey: 'AIzaSyCERb6to7cgWHC_w-3FDKF00eQZIl4oXtY',
+    appId: '1:667274875306:ios:11aa7652ecefbf2aaa9bd3',
+    messagingSenderId: '667274875306',
+    projectId: 'e-commerce-application-71a24',
+    storageBucket: 'e-commerce-application-71a24.appspot.com',
+    iosClientId:
+        '667274875306-k885j65jdfgarpv4nu2dt81bj5up9p8j.apps.googleusercontent.com',
+    iosBundleId: 'com.example.eCommerceApplication',
+  );
+
+  static const FirebaseOptions macos = FirebaseOptions(
+    apiKey: 'AIzaSyCERb6to7cgWHC_w-3FDKF00eQZIl4oXtY',
+    appId: '1:667274875306:ios:11aa7652ecefbf2aaa9bd3',
+    messagingSenderId: '667274875306',
+    projectId: 'e-commerce-application-71a24',
+    storageBucket: 'e-commerce-application-71a24.appspot.com',
+    iosClientId:
+        '667274875306-k885j65jdfgarpv4nu2dt81bj5up9p8j.apps.googleusercontent.com',
+    iosBundleId: 'com.example.eCommerceApplication',
+  );
+}
